테스트 후 수정 (#11)
* add: User Entity 생성

* update: User Entity 생성자 추가

* add: SignupRequestDto 생성. Validation 적용

* update: 회원가입 기능 구현(Bcrypt 암호화 적용)

* update: User Entity에 유저를 deactivate 하는 메서드 추가

* update: User 회원탈퇴 기능 추가

* update: User Profile 조회, 수정 기능 추가

* fix: newsfeed 명을 post로 변경

* login feature 수정중6 (isTokenValid 수정중)

* refactor: UserService.java 파일 merge

* refactor: UserService.java 파일 merge

* fix: API 수정

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken1

* RefreshToken2

* RefreshToken3

* logout 수정중1

* logout 수정중2

* logout 구현 완료

* fix: 메서드명 직관적이게 수정

* fix: 일관성 맞춰서 url 수정

* test: 기능테스트

---------

+++ src/main/java/com/sparta/newsfeedapp/dto/commentResponseDto/CommentResponseDto.java
@@ -15,7 +15,7 @@ public class CommentResponseDto {
     private Long userId;
     private Long postId;
     private String content;
-    private Long countLiked;
+    private Integer countLiked;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
